+++ app/src/main/java/com/teambeme/beme/data/remote/datasource/HomeDataSourceImpl.kt
@@ -1,10 +1,10 @@
 package com.teambeme.beme.data.remote.datasource
 
 import com.teambeme.beme.data.remote.api.HomeService
-import com.teambeme.beme.home.model.RequestModifyPublic
-import com.teambeme.beme.home.model.ResponseAnswer
-import com.teambeme.beme.home.model.ResponseAnswers
-import com.teambeme.beme.home.model.ResponseModifyData
+import com.teambeme.beme.presentation.home.model.RequestModifyPublic
+import com.teambeme.beme.presentation.home.model.ResponseAnswer
+import com.teambeme.beme.presentation.home.model.ResponseAnswers
+import com.teambeme.beme.presentation.home.model.ResponseModifyData
 import javax.inject.Inject
 
 class HomeDataSourceImpl @Inject constructor(
