+++ app/src/main/java/com/teambeme/beme/domain/repository/HomeRepository.kt
@@ -1,8 +1,8 @@
 package com.teambeme.beme.domain.repository
 
-import com.teambeme.beme.home.model.ResponseAnswer
-import com.teambeme.beme.home.model.ResponseAnswers
-import com.teambeme.beme.home.model.ResponseModifyData
+import com.teambeme.beme.presentation.home.model.ResponseAnswer
+import com.teambeme.beme.presentation.home.model.ResponseAnswers
+import com.teambeme.beme.presentation.home.model.ResponseModifyData
 
 interface HomeRepository {
     suspend fun modifyPublic(answerId: Int, publicFlag: Int): ResponseModifyData
