+++ app/src/main/java/com/teambeme/beme/presentation/home/adapter/QuestionPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.teambeme.beme.home.adapter
+package com.teambeme.beme.presentation.home.adapter
 
 import android.content.Context
 import android.content.Intent
@@ -15,13 +15,13 @@ import com.teambeme.beme.answer.view.AnswerActivity
 import com.teambeme.beme.data.local.singleton.BeMeRepository
 import com.teambeme.beme.databinding.ItemHomeMoreQuestionBinding
 import com.teambeme.beme.databinding.ItemHomeQuestionBinding
-import com.teambeme.beme.home.model.Answer
-import com.teambeme.beme.home.view.AnswerSuggestFragment
-import com.teambeme.beme.home.view.HomeFragment
-import com.teambeme.beme.home.view.InfoChangeFragment
-import com.teambeme.beme.home.view.InfoChangeFragment.InfoChangeClickListener
-import com.teambeme.beme.home.view.TransitionPublicFragment
-import com.teambeme.beme.home.viewmodel.HomeViewModel
+import com.teambeme.beme.presentation.home.model.Answer
+import com.teambeme.beme.presentation.home.view.AnswerSuggestFragment
+import com.teambeme.beme.presentation.home.view.HomeFragment
+import com.teambeme.beme.presentation.home.view.InfoChangeFragment
+import com.teambeme.beme.presentation.home.view.InfoChangeFragment.InfoChangeClickListener
+import com.teambeme.beme.presentation.home.view.TransitionPublicFragment
+import com.teambeme.beme.presentation.home.viewmodel.HomeViewModel
 import java.text.SimpleDateFormat
 import java.util.*
