+++ app/src/main/java/com/teambeme/beme/presentation/home/model/Answer.kt
@@ -1,4 +1,4 @@
-package com.teambeme.beme.home.model
+package com.teambeme.beme.presentation.home.model
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
