+++ app/src/main/java/com/teambeme/beme/presentation/home/model/ResponseQuestionData.kt
@@ -1,4 +1,4 @@
-package com.teambeme.beme.home.model
+package com.teambeme.beme.presentation.home.model
 
 import com.google.gson.annotations.SerializedName
