+++ app/src/main/java/com/teambeme/beme/presentation/home/viewmodel/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.teambeme.beme.home.viewmodel
+package com.teambeme.beme.presentation.home.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.LiveData
@@ -6,7 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.teambeme.beme.domain.repository.HomeRepository
-import com.teambeme.beme.home.model.Answer
+import com.teambeme.beme.presentation.home.model.Answer
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.launch
