+++ app/src/main/res/layout/activity_splash.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".splash.SplashActivity">
+        tools:context=".presentation.splash.SplashActivity">
 
         <ImageView
             android:layout_width="match_parent"
