+++ app/src/main/res/layout/fragment_answer_suggest.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".home.view.AnswerSuggestFragment">
+    tools:context=".presentation.home.view.AnswerSuggestFragment">
 
     <TextView
         android:id="@+id/txt_home_answer_suggestion"
