+++ app/src/main/res/layout/fragment_transition_public.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".home.view.TransitionPublicFragment">
+    tools:context=".presentation.home.view.TransitionPublicFragment">
 
     <TextView
         android:id="@+id/txt_home_transition_public"
